==> data_processing/04_statting.py <==
"""
Ingest CSV data generated by R scripts and creates a single JSON file with all stations
"""


from pathlib import Path

import ujson as json
import pandas as pd


WEEKDAYS = [
    "Montag",
    "Dienstag",
    "Mittwoch",
    "Donnerstag",
    "Freitag",
    "Samstag",
    "Sonntag",
]

target_path = Path("data/stop_stats_ranked")
target_path.mkdir(exist_ok=True)


def main():
    df = pd.read_csv("data/stats_long_corrected.csv")

    df_stops = (
        df.groupby(by=["stop_name"], as_index=False)
        .aggregate(
            {
                "municipality": "first",
                "dep_per_hour_avg": "first",
                "lat": "first",
                "lon": "first",
            }
        )
        .sort_values(by=["dep_per_hour_avg"], ascending=False)
    )

    stops = list(zip(df_stops["stop_name"], df_stops["municipality"]))
    stops_with_coords = list(
        zip(
            df_stops["stop_name"],
            df_stops["municipality"],
            df_stops["lat"],
            df_stops["lon"],
        )
    )

    with open("data/stops.json", "w", encoding="utf-8") as f:
        json.dump(stops, f, ensure_ascii=False)

    with open("data/stops_with_coords.json", "w", encoding="utf-8") as f:
        json.dump(stops_with_coords, f, ensure_ascii=False)


if __name__ == "__main__":
    main()
